Move `Response` class to `React\Http\Message\Response`
