+++ examples/72-server-http-connect-proxy.php
@@ -5,7 +5,7 @@
 
 use Psr\Http\Message\ServerRequestInterface;
 use React\EventLoop\Factory;
-use React\Http\Response;
+use React\Http\Message\Response;
 use React\Http\Server;
 use React\Socket\Connector;
 use React\Socket\ConnectionInterface;
