+++ examples/81-server-upgrade-echo.php
@@ -19,7 +19,7 @@ $ telnet localhost 1080
 
 use Psr\Http\Message\ServerRequestInterface;
 use React\EventLoop\Factory;
-use React\Http\Response;
+use React\Http\Message\Response;
 use React\Http\Server;
 use React\Stream\ThroughStream;
