+++ tests/Io/StreamingServerTest.php
@@ -5,7 +5,7 @@ namespace React\Tests\Http\Io;
 use Psr\Http\Message\ServerRequestInterface;
 use React\Http\Io\StreamingServer;
 use React\EventLoop\Factory;
-use React\Http\Response;
+use React\Http\Message\Response;
 use React\Promise\Promise;
 use React\Stream\ThroughStream;
 use React\Tests\Http\SocketServerStub;
