+++ tests/Middleware/StreamingRequestMiddlewareTest.php
@@ -4,7 +4,7 @@ namespace React\Tests\Http\Middleware;
 
 use React\Http\Middleware\StreamingRequestMiddleware;
 use React\Http\Io\ServerRequest;
-use React\Http\Response;
+use React\Http\Message\Response;
 use React\Tests\Http\TestCase;
 
 class StreamingRequestMiddlewareTest extends TestCase
